Refactor: 프로젝트 폴더 구조 변경 및 불필요한 .gitkeep 파일 삭제
