+++ src/main/java/com/chingubackend/jwt/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.chingubackend.security;
+package com.chingubackend.jwt;
 
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
